Move healt check to internal package
