+++ Sources/HealthCheck/DateFormatters.swift
@@ -1,3 +1,5 @@
+// Date formatters for use in Health API
+// Enables us to output iso8601 formatted dates in our API
 import Foundation
 
 struct DateFormatters {
@@ -5,13 +7,6 @@ struct DateFormatters {
     public static var iso8601: DateFormatter {
         let formatter = DateFormatter()
         formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZ"
-        return formatter
-    }
-
-    // yyyy-MM-dd
-    public static var date: DateFormatter {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
         return formatter
     }
 }
